@@ -6,7 +6,7 @@ shared_examples_for 'maps JSON fields' do |mapped_attributes|
 
     mapped_attributes.each do |from, to|
       it "maps #{from.inspect} to #{to.inspect}" do
-        subject[to].should == json[from]
+        expect(subject[to]).to eq(json[from])
       end
     end
   end
@@ -18,7 +18,7 @@ shared_examples_for 'ignores JSON fields' do |ignored_attributes|
   describe described_class.send(:json) do
     ignored_attributes.each do |name|
       it "ignores #{name.inspect}" do
-        json_handler.new(json.merge(name => 'ignored')).should_not have_key(name)
+        expect(json_handler.new(json.merge(name => 'ignored'))).to_not have_key(name)
       end
     end
   end
